third_party/blink/web_tests/fast/dom/Element/parent-node-interface-expected.txt

This tests that Element implements the ParentNode interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS element.children.length is 0
PASS element.childElementCount is 0
PASS element.firstElementChild is null
PASS element.lastElementChild is null
PASS element.children.length is 0
PASS element.childElementCount is 0
PASS element.firstElementChild is null
PASS element.lastElementChild is null
PASS element.children.length is 1
PASS element.childElementCount is 1
PASS element.children[0] is b
PASS element.firstElementChild is b
PASS element.lastElementChild is b
PASS element.children.length is 1
PASS element.childElementCount is 1
PASS element.children[0] is b
PASS element.firstElementChild is b
PASS element.lastElementChild is b
PASS element.children.length is 2
PASS element.childElementCount is 2
PASS element.children[0] is b
PASS element.children[1] is d
PASS element.firstElementChild is b
PASS element.lastElementChild is d
PASS element.children.length is 2
PASS element.childElementCount is 2
PASS element.children[0] is b
PASS element.children[1] is d
PASS element.firstElementChild is b
PASS element.lastElementChild is d
PASS successfullyParsed is true

TEST COMPLETE
